Tests: split Codex hook timeout regression support
